add basic hero movement

+++ src/main/java/com/bombasticoctocat/bomberman/game/Particle.java
@@ -4,8 +4,8 @@ package com.bombasticoctocat.bomberman.game;
  * Created by kustosz on 04/05/14.
  */
 public interface Particle {
-    public int getX();
-    public int getY();
+    public double getX();
+    public double getY();
     public int width();
     public int height();
 }
